requirements/dev.txt
Django>=1.10.3
django-tinymce>=2.3.0
django-debug-toolbar>=1.5
psycopg2>=2.6.1
unidecode>=0.4.20
Pillow>=3.2.0
django-simple-captcha>=0.5.5
